+++ CHANGES.md
@@ -1,1 +1,2 @@
+* 2014-04-14 - Added Candidate.only_permit_levels_up_to to allow the maximum acceptable values to be set broadly
 * 2014-04-14 - Candidate now has class methods that allow acceptable values to be set easily (e.g., Candidate.nuclear_power = [1] prevents any nuclear power being chosen).